Centralise configuration; remove Stub
Previously, the Example app had two ways of getting config. There was the stub way, and the remote way. It also cached them for 1 day.

We've remove the two ways. It now always loads it from the back end, or uses the cache. If it can't get that, it loads has a fall-back default config to use.

This means the app has a little less user-interface. There's no longer an idea of the 'stub' configuration. A QA will need to edit the config by changing the response from the server.