Allow editing all options on cart screen
Previously, we could only edit the “BuyNow” option on the cart screen, before checkout.

Now, we have switches for all three options. The logic controller maintains the source of truth for the options. Event handlers update this source of source of truth via the logic controller’s `toggleOption(:)` function. That function takes a CheckoutV2Options key path, and toggles it.