Add features struct, with widget enabled feature
Previously, there was no way to have feature flags in the SDK. Now, we are using the defaults parsed from the app’s arguments on launch.

With these arguments are passed on app launch: `-com.afterpay.widget-enabled YES`, the SDK will have the `widgetEnabled` feature toggle turned on. If it’s not passed, it is off by default.

Currently, no behaviour is based on this feature flag. Soon, it will be used to turn on the checkout widget.