Add simple read-only WidgetView
This WidgetView does not do much other than host a web view. It’s a web view inside a UIKit view. That’s pretty much all for now, but it will be doing a lot more soon. For now, it will only load the Afterpay widget via the bootstrap page.

The bootstrap is being hosted and developed elsewhere. Suffice to say, it requires someone to call the javascript function `createAfterpayWidget` with a successful checkout token. This is what the WidgetView currently does.

It does not:

- update in any way
- inform us of any changes it makes itself
- support proper intrinsic sizing
- have any kind of nice boundary between the iOS view and web view.