Add WidgetView to MessageViewController
Previously, the MessageViewController just had a message on it. Usually it was a happy “Success” message, and that was it.

Now, we’ll be using it to demonstrate the WidgetView. It’ll have one of those, too (if the feature is enabled).

Because the WidgetView doesn’t do much right now, it’s just plopped right onto the screen. It’s got a height constraint of 350 because the WidgetView doesn’t know how to size itself correctly yet. It will one day.