Remove script handler on view dismissal
While loading the view, we make the CheckoutV2ViewController its own message handler with: `userContentController.add(self, name: "iOS”)`. This retains the view controller, and causes a memory leak.

We should remove the script message handler when its no longer needed. Since we setup the handler in `loadView`, we should remove it in `viewWillDisappear`.