Receive and handle events from widget
Previously, we would set up the web widget, and that would be that. Now, when we provide Afterpay with a `WidgetHandler`, that `WidgetHandler` will be sent any events from the web widget.

In order to this, we have made our `WidgetView` into a `WKScriptMessageHandler`. It sets itself as a handler for a user content controller. It receives events from the web widget via the `WKScriptMessageHandler` call-back method. All widget events go through this. It decodes them (they are typed as `WidgetEvent` in Swift). Depending on the kind of event, it calls the appropriate `WidgetHandler` method.

We have mapped the `WidgetHandler` methods directly to their equivalent on the web. It doesn’t have to be that way, but this is the case for now.

The `WidgetEvent` enum attempts to model all the different kind of events that can happen to the web widget.

TODO: Documentation!