Add `onFailure` callback to `WidgetHandler`
This call back is called when something has happened inside the SDK from which we cannot recover. For example, if there is a JSON decoding issue with the events from the web widget.

This is necessary because we do not want the web widget and the merchant app checkout screen to get out of sync.  A change may happen in the web widget that does not get propagated to the merchant app because of some error in the SDK. In that situation, we need to let the merchant app know. If we don't, the web widget may be displaying something that is in conflict with the merchant app.