Add method to get the widget's status
Previously, we only had the WidgetHandler call-backs, which would be called when the web widget changed status. This method can be called by the merchant app can call this method any time, and receive the current status in the method's callback.

There are several JavaScript properties on the web widget which can be used to ask about the widget's status. Rather than calling them one by one, decoding, and sending back the value in a call-back, we want to merge them together in one status value. Fortunately, we already have a structure which fits this bill: the WidgetStatus.

The bootstrap pages' JavaScript is responsible for merging the web widget's JavaScript properties together into one object. The object which it returns is decodable into a WidgetStatus.