Get WidgetView’s intrinsic size from web view
Previously, we had to constrain the `WidgetView` to a fixed height. Now, it has an intrinsic content size which comes from the web view. That means the web view controls how big to make the `WidgetView`.

The web view does occasionally need to signal to us that it needs to change size, such as after loading. We’ve introduced a new `WidgetEvent` for this: `resize`. We use a MutationObserver in the bootstrap to send the resize via postMessage. On the iOS side, we handle a resize by calling `invalidateIntrinsicContentSize`. We don’t need to inform the `WidgetHandler` about a resize.

Since we made `WidgetEvent` conform to `Equatable`, we also had to make the `WidgetStatus` and `Money` types `Equatable` as well.