Add token-less option to WidgetView
Previously, we had to initialise a widget with some token. Now, we want a token-less version. This will be used _before_ a checkout, so when a token is not known but an amount is.

We've added an initialiser for this purpose. It just takes an amount of money in the same currency that Afterpay is configured to use (it throws an exception if there is no configuration yet).

If configured for token-less, it sends just the amount over to the web widget's bootstrap page.