Get locale from the configuration
Previously, the widget bootstrap always hard-coded the `en-US` locale. That was fine for developing, but now the widget supports a lot more locales.

We need to pass the locale in from the iOS app. It's passed in when calling the `createAfterpayWidget` function.

The Widget only supposed locales which use `-`s in their names: `en-US` and `en-UK`, etc. iOS tends to send them with underscores `_`: eg. `en_US`, `en_UK`. To prevent confusion, we accept both. The bootstrap simply converts `_`s to `-`s.