Send suggested height with resize event
The bootstrap JS can read the suggested size from the `offsetHeight`. We have also added 16 px above and below the `#afterpay-widget-container` to give us more room. The height then gets sent to the iOS app via the existing `resize` event, but it also contains a `size`.

When the `WidgetView` on iOS receives the resize event, with a size, it uses that height as the widget view’s intrinsic content size height. If the height is less than 250, it uses 250. The minimum height of the WidgetView is therefore 250.