Send style options to the web widget
Previously, we had hard-coded the widget’s style object in the bootstrap. Now, we pass those options from the iOS WidgetView. These can be customised by the merchant when they initialise the WidgetView.

The WidgetView is initialised with a Style struct which has the two styling options on it, logo and heading. They are both on by default. We encode that and send it to the bootstrap when we create the widget.