Inject afterpay.js script element from iOS
Previously, the afterpay.js script element was simply part of the bootstrap. This meant you couldn’t change the URL the script was coming from.

Now, we are injecting the script tag when we setup the web view. Shortly after it’s injected, we run the code to set up the widget.