Add `CheckoutMode` for checkout request
The checkouts request structure sends along a `mode` property. Previously, it was hard-wired to always send it as ‘express’. This means the checkouts API would think we were always asking for an express checkout. For redirect checkout (aka v1), we will not want the checkouts API to think that. In that instance, we want to not send the `mode` at all.

We introduce a `CheckoutMode` enum, with both v1 and v2 cases. When using the checkout APIClient function, we pass along the mode we want to use. If it’s v2, we send the value `express`. If it’s v1, we send along nothing.