Add v1 checkout (non-express checkout)
Previously, opening the checkout would always give you the express checkout. We’ve added a switch to the CheckoutOptionsCell which turns off the express checkout, and that allows you to open the regular v1 checkout.

The switch for express works in the same way as the switch for the other checkout options. When tapped, it calls the event handler with the `expressToggled` event. The flow controller receives that, and tells the logic controller about it via calling `toggleExpressCheckout`.

Ultimately, when it’s time to pay, the logic controller then knows whether it should call the express checkout or whether it should call the v1 checkout. We’ve added a new `Command` to the logic controller’s command enum to handle checking out with v1. Checkout v1 doesn’t need options. Instead it needs a URL. We fetch the URL then send as an associated value along with the command.

We also disable the checkout v2 options when the express switch is off — this helps the user understand that those options only apply to express checkout.