Remove unused textfields (and fix crash...)
These text fields were made defunct in 88caa3c7e7853527b79774618479922589cf0e16, but I accidentally left their implicitly-unwrapped optional properties around.

When the keyboard resized, and the notification center called the `adjustForKeyboard` method, it referred to the properties. Since they were implicitly unwrapped, and never had a value assigned, this was a crash.

We can simply remove the properties, and remove the part of the `adjustForKeyboard` method which moved the scroll view to be aligned with the text fields.