Do not inject afterpay.js script tag
Previously, we were injecting the script tag via the WidgetView. This meant the WidgetView could decide the URL from which we could get afterpay.js. This turned out to not be that advantageous because staging and prod are rarely meaningfully different from our point of view.

Instead, we’ll move the script tag back into the bootstrap html, but we’ll put it in the header this time. That way it will be ready by the time we can execute the rest of the scripts. This will speed things up for us, (and make it work on Android!)