mv MessageViewController to WidgetViewController
Previously, we had a `MessageViewController` which had the only job of showing simple text messages to the user. It no longer does that - its job is instead to show the Widget. Renaming it will make more sense.

We’ve also changed it’s initialisers to be either a token, or a money amount. This reflects the two modes in which it can show a widget: token mode, or tokenless (money) mode.