Make super class init available internally
Previously, we had marked some NSObject superclasses’ inits as unavailable. That way, people won’t accidentally instantiate them:

```
  class Super: NSObject {
    @available(*, unavailable)
    override init() { }
  }

  class Sub: Super {
    let s: String
    init(s: String) { self.s = s }
  }
```

However, in Swift 5.4, this changed. The subclasses’ inits do not compile because they cannot call the super.

Instead, we mark the super classes’ initialiser as `internal` so that consumers of the Afterpay SDK cannot accidentally initialise the super.