Embed widget bootstrap HTML inside app
Previously, we were loading the bootstrap HTML + JS from GitHub pages. This was okay, except when we wanted to change environments between staging and production. The URL for Afterpay.js needs to change, and since the HTML was centrally hosted, there was no way to change it.

Instead, we can host just the bootstrap’s JS centrally on GitHub, and keep the HTML inside the app. This way, we can adjust the URLs it needs depending on the environment.

A different call is used to start the web view. Instead of `webView.load`, we use `webView.loadHTMLString(:baseURL:)`. In there we pass the HTML, but we also give the widget bootstrap URL to use as a `baseURL`. That prevents cross site scripting issues occurring inside the webview.