Remove disk caches for bootstrap
Previously, the checkout view v2 controller removed its disk cache for its bootstrap. This way, it was always fetching a new bootstrap. In the checkout, we had set its cache policy to never cache, so that it was always getting a fresh bootstrap too. When we changed the checkout to load HTML that is stored in the app, we no longer get the option to set the caching policy.

Now, I have moved the disk cache clearing code to a central place. We call it for both the checkout v2 and the widget.