Re-enable widget UI test with mock bootstrap
Previously, we had disabled the widget UI test. Now, instead of loading the real widget via the real bootstrap, we use a mock bootstrap.

The mock bootstrap works by replacing the `afterpay-widget-container` with a message. It does this in response to the iOS WidgetView calling its methods. We can test that message.