Show cancel alert with `.alert` style on iPad
Previously, we showed the 'are you sure you want to cancel the payment' alert with the preferred style of '.actionSheet'. This looks good on iPhone, but on iPad, it hits an assertion failure in UIKit. On iPad, an `.actionSheet` style needs a sourceView and sourceRect or a barButtonItem.

Such an alert style would look out of place, anyway, so we have changed the alert style to be `.alert` on iPad and `.actionSheet` on iPhone.